+++ src/main/java/seedu/address/logic/commands/cheatsheetcommands/AddCheatSheetCommand.java
@@ -1,8 +1,8 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.cheatsheetcommands;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.ADD;
-import static seedu.address.logic.commands.EditCheatSheetCommand.createEditedCheatSheet;
+import static seedu.address.logic.commands.cheatsheetcommands.EditCheatSheetCommand.createEditedCheatSheet;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TITLE;
 
@@ -11,6 +11,8 @@ import java.util.Set;
 
 import javafx.collections.ObservableList;
 
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.cheatsheet.CheatSheet;
