+++ src/main/java/seedu/address/logic/commands/cheatsheetcommands/DeleteCheatSheetCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.cheatsheetcommands;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.DELETE;
@@ -7,6 +7,8 @@ import java.util.List;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.cheatsheet.CheatSheet;
