+++ src/main/java/seedu/address/logic/parser/AddNoteCommandParser.java
@@ -8,7 +8,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TITLE;
 import java.util.Set;
 import java.util.stream.Stream;
 
-import seedu.address.logic.commands.AddNoteCommand;
+import seedu.address.logic.commands.notecommands.AddNoteCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.note.Content;
 import seedu.address.model.note.Note;
