+++ src/main/java/seedu/address/logic/parser/DeleteFlashcardCommandParser.java
@@ -3,7 +3,7 @@ package seedu.address.logic.parser;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.DeleteFlashcardCommand;
+import seedu.address.logic.commands.flashcardcommands.DeleteFlashcardCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
