+++ src/main/java/seedu/address/logic/parser/FilterCheatSheetByTagCommandParser.java
@@ -5,7 +5,7 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT
 import java.util.HashSet;
 import java.util.Set;
 
-import seedu.address.logic.commands.FilterCheatSheetByTagCommand;
+import seedu.address.logic.commands.cheatsheetcommands.FilterCheatSheetByTagCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.cheatsheet.CheatSheetContainsTagPredicate;
 import seedu.address.model.tag.Tag;
