+++ src/main/java/seedu/address/logic/parser/FilterFlashcardByTagCommandParser.java
@@ -5,7 +5,7 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT
 import java.util.HashSet;
 import java.util.Set;
 
-import seedu.address.logic.commands.FilterFlashcardByTagCommand;
+import seedu.address.logic.commands.flashcardcommands.FilterFlashcardByTagCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.flashcard.FlashcardContainsTagPredicate;
 import seedu.address.model.tag.Tag;
